+++ semcore/tag/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.4.35] - 2023-05-22
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [3.50.7 ~> 3.51.0], `@semcore/icon` [3.14.11 ~> 3.14.12], `@semcore/flex-box` [4.7.23 ~> 4.7.24]).
+
+## [4.4.34] - 2023-05-11
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [3.50.6 ~> 3.50.7], `@semcore/icon` [3.14.10 ~> 3.14.11], `@semcore/flex-box` [4.7.22 ~> 4.7.23]).
+
 ## [4.4.33] - 2023-05-04
 
 ### Changed